refactor!: simplify evaluation_context namespace
