+++ open_feature/hooks/hook_context.py
@@ -1,7 +1,7 @@
 import typing
 from dataclasses import dataclass
 
-from open_feature.evaluation_context.evaluation_context import EvaluationContext
+from open_feature.evaluation_context import EvaluationContext
 from open_feature.flag_evaluation.flag_type import FlagType
