+++ open_feature/hooks/hook_support.py
@@ -2,7 +2,7 @@ import logging
 import typing
 from functools import reduce
 
-from open_feature.evaluation_context.evaluation_context import EvaluationContext
+from open_feature.evaluation_context import EvaluationContext
 from open_feature.flag_evaluation.flag_evaluation_details import FlagEvaluationDetails
 from open_feature.flag_evaluation.flag_type import FlagType
 from open_feature.hooks.hook import Hook
