+++ open_feature/open_feature_api.py
@@ -1,6 +1,6 @@
 import typing
 
-from open_feature.evaluation_context.evaluation_context import EvaluationContext
+from open_feature.evaluation_context import EvaluationContext
 from open_feature.exception.exceptions import GeneralError
 from open_feature.hooks.hook import Hook
 from open_feature.open_feature_client import OpenFeatureClient
