+++ open_feature/open_feature_client.py
@@ -3,7 +3,7 @@ import typing
 from dataclasses import dataclass
 
 from open_feature import open_feature_api as api
-from open_feature.evaluation_context.evaluation_context import EvaluationContext
+from open_feature.evaluation_context import EvaluationContext
 from open_feature.exception.error_code import ErrorCode
 from open_feature.exception.exceptions import (
     GeneralError,
