+++ open_feature/provider/in_memory_provider.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 import typing
 
 from open_feature._backports.strenum import StrEnum
-from open_feature.evaluation_context.evaluation_context import EvaluationContext
+from open_feature.evaluation_context import EvaluationContext
 from open_feature.exception.error_code import ErrorCode
 from open_feature.flag_evaluation.reason import Reason
 from open_feature.flag_evaluation.resolution_details import FlagResolutionDetails
