+++ open_feature/provider/provider.py
@@ -1,7 +1,7 @@
 import typing
 from abc import abstractmethod
 
-from open_feature.evaluation_context.evaluation_context import EvaluationContext
+from open_feature.evaluation_context import EvaluationContext
 from open_feature.flag_evaluation.resolution_details import FlagResolutionDetails
 from open_feature.hooks.hook import Hook
 from open_feature.provider.metadata import Metadata
